pyproject.toml
[tool.poetry]
name = "e-trading-network"
version = "0.1.0"
description = "test work"
authors = ["Yuriy S. <trupakenator@gmail.com>"]
readme = "README.md"
packages = [{include = "e_trading_network"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
psycopg2-binary = "^2.9.5"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
